pyproject.toml
[tool.poetry]
name = "flaskwebapp"
version = "0.1.0"
description = "FlaskWebAppDevelopemnt in Ubuntu"
authors = ["Ram"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"
jinja2 = "^3.1.4"
uvicorn = "^0.30.1"
asgiref = "^3.8.1"
flask-smorest = "^0.44.0"
flask-sqlalchemy = "^3.1.1"
flask-jwt-extended = "^4.6.0"
passlib = "^1.7.4"
flask-migrate = "^4.0.7"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"
autopep8 = "^2.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
